Use Spring WebFlux and WebClient
Spring Framework now ships a new reactive HTTP client in the WebFlux
module.

This commit replaces the `RestTemplateBuilder` with an auto-configured
`WebClient.Builder`. This client leverages the reactive API `Mono` and
`Flux`.

We're not really doing anything extra here as the calls toe the client
are blocked so whatever performance problem we have should still be
around.